+++ example/android/app/src/main/kotlin/com/example/test_animate_icon/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.test_animate_icon
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}