+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 * First Working release to Animate Icons
 
-
 ## [0.0.2] - 21/06/2020
 
 * Update Gif to Links 
+
+## [0.0.3] - 21/06/2020
+
+* Added an example to use the pakage and updated description 
+
+## [0.0.4] - 05/07/2020
+
+* Added AniamteIconController to animate without onPress. 
+* The onStartIconPress, onEndIconPress need to return a bool value, if the value returned is true, than the animation will happen and if it returns false, then the animation will not happen.